Moved `StatesSavedReadCmd` and `StatesDesiredReadCmd` up one level.
