+++ crate/rt/src/cmds.rs
@@ -10,8 +10,9 @@
 
 pub use self::{
     clean_cmd::CleanCmd, cmd_base::CmdBase, diff_cmd::DiffCmd, ensure_cmd::EnsureCmd,
-    states_desired_display_cmd::StatesDesiredDisplayCmd, states_discover_cmd::StatesDiscoverCmd,
-    states_saved_display_cmd::StatesSavedDisplayCmd,
+    states_desired_display_cmd::StatesDesiredDisplayCmd,
+    states_desired_read_cmd::StatesDesiredReadCmd, states_discover_cmd::StatesDiscoverCmd,
+    states_saved_display_cmd::StatesSavedDisplayCmd, states_saved_read_cmd::StatesSavedReadCmd,
 };
 
 pub mod cmd_ctx_internal;
@@ -22,5 +23,7 @@ mod cmd_base;
 mod diff_cmd;
 mod ensure_cmd;
 mod states_desired_display_cmd;
+mod states_desired_read_cmd;
 mod states_discover_cmd;
 mod states_saved_display_cmd;
+mod states_saved_read_cmd;
