+++ examples/envman/src/cmds/env_clean_cmd.rs
@@ -2,7 +2,7 @@ use futures::FutureExt;
 use peace::{
     cmd::scopes::{SingleProfileSingleFlowView, SingleProfileSingleFlowViewAndOutput},
     fmt::presentable::{Heading, HeadingLevel, ListNumbered},
-    rt::cmds::{sub::StatesSavedReadCmd, CleanCmd},
+    rt::cmds::{CleanCmd, StatesSavedReadCmd},
     rt_model::{outcomes::CmdOutcome, output::OutputWrite},
 };
