+++ examples/envman/src/cmds/env_deploy_cmd.rs
@@ -2,7 +2,7 @@ use futures::FutureExt;
 use peace::{
     cmd::scopes::{SingleProfileSingleFlowView, SingleProfileSingleFlowViewAndOutput},
     fmt::presentable::{Heading, HeadingLevel, ListNumbered},
-    rt::cmds::{sub::StatesSavedReadCmd, EnsureCmd},
+    rt::cmds::{EnsureCmd, StatesSavedReadCmd},
     rt_model::{outcomes::CmdOutcome, output::OutputWrite},
 };
