+++ workspace_tests/src/rt/cmds/clean_cmd.rs
@@ -2,7 +2,7 @@ use peace::{
     cfg::{app_name, profile, AppName, FlowId, Profile},
     cmd::ctx::CmdCtx,
     resources::states::StatesSaved,
-    rt::cmds::{sub::StatesSavedReadCmd, CleanCmd, EnsureCmd, StatesDiscoverCmd},
+    rt::cmds::{CleanCmd, EnsureCmd, StatesDiscoverCmd, StatesSavedReadCmd},
     rt_model::{outcomes::CmdOutcome, Flow, ItemGraphBuilder, Workspace, WorkspaceSpec},
 };
