+++ workspace_tests/src/rt/cmds/ensure_cmd.rs
@@ -2,7 +2,7 @@ use peace::{
     cfg::{app_name, profile, AppName, FlowId, Profile},
     cmd::ctx::CmdCtx,
     resources::states::StatesSaved,
-    rt::cmds::{sub::StatesSavedReadCmd, EnsureCmd, StatesDiscoverCmd},
+    rt::cmds::{EnsureCmd, StatesDiscoverCmd, StatesSavedReadCmd},
     rt_model::{outcomes::CmdOutcome, Flow, ItemGraphBuilder, Workspace, WorkspaceSpec},
 };
